# Speed up aws plugin load