Using ```brew --cellar``` instead of ```brew list``` to find the awscli path avoids about 166 shell commands and a ruby binary invocation on every zsh startup. 

The aws plugin currently adds 0.63 seconds to my zshrc startup, which accounts for roughly 65% of my zsh's startup time. After making this change, this has been reduced to 9% of startup time. The 0.54 second speed up may not seem like a lot, but it adds up in even a modestly sized zshrc setup, or one with a lot of omz plugins.

brew list:
```
+ set -x
+ PS4='+\t '
+14:54:40 set +o posix
++14:54:40 quiet_cd /usr/local/bin/
++14:54:40 cd /usr/local/bin/
++14:54:40 pwd -P
+14:54:40 BREW_FILE_DIRECTORY=/usr/local/bin
+14:54:40 HOMEBREW_BREW_FILE=/usr/local/bin/brew
+14:54:40 HOMEBREW_PREFIX=/usr/local
+14:54:40 [[ -z /usr/local ]]
+14:54:40 [[ /usr/local = \/\u\s\r\/\l\o\c\a\l\/\b\i\n\/\b\r\e\w ]]
+14:54:40 HOMEBREW_REPOSITORY=/usr/local
+14:54:40 [[ -L /usr/local/bin/brew ]]
++14:54:40 symlink_target_directory /usr/local/bin/brew /usr/local/bin
+++14:54:40 readlink /usr/local/bin/brew
++14:54:40 local target=../Homebrew/bin/brew
+++14:54:40 dirname ../Homebrew/bin/brew
++14:54:40 local target_dirname=../Homebrew/bin
++14:54:40 local directory=/usr/local/bin
++14:54:40 quiet_cd /usr/local/bin
++14:54:40 cd /usr/local/bin
++14:54:40 quiet_cd ../Homebrew/bin
++14:54:40 cd ../Homebrew/bin
++14:54:40 pwd -P
+14:54:40 BREW_FILE_DIRECTORY=/usr/local/Homebrew/bin
+14:54:40 HOMEBREW_REPOSITORY=/usr/local/Homebrew
+14:54:40 [[ -L /usr/local/bin/brew ]]
++14:54:40 symlink_target_directory /usr/local/bin/brew /usr/local/bin
+++14:54:40 readlink /usr/local/bin/brew
++14:54:40 local target=../Homebrew/bin/brew
+++14:54:40 dirname ../Homebrew/bin/brew
++14:54:40 local target_dirname=../Homebrew/bin
++14:54:40 local directory=/usr/local/bin
++14:54:40 quiet_cd /usr/local/bin
++14:54:40 cd /usr/local/bin
++14:54:40 quiet_cd ../Homebrew/bin
++14:54:40 cd ../Homebrew/bin
++14:54:40 pwd -P
+14:54:40 USR_LOCAL_BREW_FILE_DIRECTORY=/usr/local/Homebrew/bin
+14:54:40 USR_LOCAL_HOMEBREW_REPOSITORY=/usr/local/Homebrew
+14:54:40 [[ /usr/local/Homebrew = \/\u\s\r\/\l\o\c\a\l\/\H\o\m\e\b\r\e\w ]]
+14:54:40 HOMEBREW_PREFIX=/usr/local
+14:54:40 HOMEBREW_LIBRARY=/usr/local/Homebrew/Library
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z vim ]]
+14:54:40 VAR_NEW=HOMEBREW_EDITOR
+14:54:40 [[ -n '' ]]
+14:54:40 export HOMEBREW_EDITOR=vim
+14:54:40 HOMEBREW_EDITOR=vim
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 VAR_NEW=HOMEBREW_PATH
+14:54:40 [[ -n '' ]]
+14:54:40 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:54:40 [[ -z '' ]]
+14:54:40 continue
+14:54:40 [[ -n '' ]]
+14:54:40 source /usr/local/Homebrew/Library/Homebrew/brew.sh
+++14:54:40 git -C /usr/local/Homebrew describe --tags --dirty
++14:54:40 HOMEBREW_VERSION=1.2.1-96-gfb33acb-dirty
++14:54:40 HOMEBREW_USER_AGENT_VERSION=1.2.1-96-gfb33acb-dirty
++14:54:40 [[ -z 1.2.1-96-gfb33acb-dirty ]]
++14:54:40 export HOMEBREW_COMMAND_DEPTH=1
++14:54:40 HOMEBREW_COMMAND_DEPTH=1
+++14:54:40 locale charmap
++14:54:40 [[ UTF-8 != \U\T\F\-\8 ]]
++14:54:40 [[ -d /usr/local/Homebrew/Cellar ]]
++14:54:40 HOMEBREW_CELLAR=/usr/local/Cellar
++14:54:40 case "$*" in
++14:54:40 [[ /usr/local = \/ ]]
++14:54:40 [[ /usr/local = \/\u\s\r ]]
++14:54:40 export GEM_OLD_HOME=
++14:54:40 GEM_OLD_HOME=
++14:54:40 export GEM_OLD_PATH=
++14:54:40 GEM_OLD_PATH=
++14:54:40 unset GEM_HOME
++14:54:40 unset GEM_PATH
++14:54:40 unset BASH_ENV
++14:54:40 unset GREP_OPTIONS
+++14:54:40 uname -s
++14:54:40 HOMEBREW_SYSTEM=Darwin
++14:54:40 case "$HOMEBREW_SYSTEM" in
++14:54:40 HOMEBREW_MACOS=1
++14:54:40 HOMEBREW_CURL=/usr/bin/curl
++14:54:40 [[ -n 1 ]]
+++14:54:40 uname -p
++14:54:40 HOMEBREW_PROCESSOR=i386
++14:54:40 HOMEBREW_PRODUCT=Homebrew
++14:54:40 HOMEBREW_SYSTEM=Macintosh
++14:54:40 [[ i386 = \i\3\8\6 ]]
++14:54:40 HOMEBREW_PROCESSOR=Intel
+++14:54:40 /usr/bin/sw_vers -productVersion
++14:54:40 HOMEBREW_MACOS_VERSION=10.11.6
++14:54:40 HOMEBREW_OS_VERSION='macOS 10.11.6'
++14:54:40 HOMEBREW_OS_USER_AGENT_VERSION='Mac OS X 10.11.6'
++14:54:40 printf -v HOMEBREW_MACOS_VERSION_NUMERIC %02d%02d%02d 10 11 6
++14:54:40 [[ 101106 -lt 100900 ]]
++14:54:40 HOMEBREW_USER_AGENT='Homebrew/1.2.1-96-gfb33acb-dirty (Macintosh; Intel Mac OS X 10.11.6)'
+++14:54:40 /usr/bin/curl --version
+++14:54:40 head -n1
+++14:54:40 /usr/bin/awk '{print $1"/"$2}'
++14:54:40 HOMEBREW_CURL_VERSION=curl/7.43.0
++14:54:40 HOMEBREW_USER_AGENT_CURL='Homebrew/1.2.1-96-gfb33acb-dirty (Macintosh; Intel Mac OS X 10.11.6) curl/7.43.0'
++14:54:40 [[ -z '' ]]
++14:54:40 export HOMEBREW_BREW_FILE
++14:54:40 export HOMEBREW_PREFIX
++14:54:40 export HOMEBREW_REPOSITORY
++14:54:40 export HOMEBREW_LIBRARY
++14:54:40 export HOMEBREW_VERSION
++14:54:40 export HOMEBREW_CACHE
++14:54:40 export HOMEBREW_CELLAR
++14:54:40 export HOMEBREW_SYSTEM
++14:54:40 export HOMEBREW_CURL
++14:54:40 export HOMEBREW_PROCESSOR
++14:54:40 export HOMEBREW_PRODUCT
++14:54:40 export HOMEBREW_OS_VERSION
++14:54:40 export HOMEBREW_MACOS_VERSION
++14:54:40 export HOMEBREW_USER_AGENT
++14:54:40 export HOMEBREW_USER_AGENT_CURL
++14:54:40 [[ -n 1 ]]
++14:54:40 [[ -x /usr/bin/xcode-select ]]
+++14:54:40 /usr/bin/xcode-select --print-path
++14:54:40 XCODE_SELECT_PATH=/Library/Developer/CommandLineTools
++14:54:40 [[ /Library/Developer/CommandLineTools = \/ ]]
++14:54:40 [[ -n /Library/Developer/CommandLineTools ]]
+++14:54:40 /usr/bin/xcrun clang
++14:54:40 XCRUN_OUTPUT='clang: error: no input files'
++14:54:40 XCRUN_STATUS=1
++14:54:40 [[ 1 -ne 0 ]]
++14:54:40 [[ clang: error: no input files = *license* ]]
+++14:54:40 pwd
++14:54:40 [[ list = -v ]]
++14:54:40 for arg in '"$@"'
++14:54:40 [[ list = \-\-\h\e\l\p ]]
++14:54:40 [[ list = \-\h ]]
++14:54:40 [[ list = \-\-\u\s\a\g\e ]]
++14:54:40 [[ list = \-\? ]]
++14:54:40 HOMEBREW_ARG_COUNT=1
++14:54:40 HOMEBREW_COMMAND=list
++14:54:40 shift
++14:54:40 case "$HOMEBREW_COMMAND" in
++14:54:40 [[ -z '' ]]
++14:54:40 export HOMEBREW_GIT_CONFIG_FILE=/usr/local/Homebrew/.git/config
++14:54:40 HOMEBREW_GIT_CONFIG_FILE=/usr/local/Homebrew/.git/config
+++14:54:40 git config --file=/usr/local/Homebrew/.git/config --get homebrew.devcmdrun
++14:54:40 HOMEBREW_GIT_CONFIG_DEVELOPERMODE=true
++14:54:40 [[ true = \t\r\u\e ]]
++14:54:40 export HOMEBREW_DEV_CMD_RUN=1
++14:54:40 HOMEBREW_DEV_CMD_RUN=1
++14:54:40 [[ -f /usr/local/Homebrew/Library/Homebrew/cmd/list.sh ]]
++14:54:40 [[ -f /usr/local/Homebrew/Library/Homebrew/dev-cmd/list.sh ]]
++14:54:40 check-run-command-as-root
+++14:54:40 id -u
++14:54:40 [[ 502 = 0 ]]
++14:54:40 return
++14:54:40 [[ /usr/local = \/\u\s\r\/\l\o\c\a\l ]]
++14:54:40 [[ /usr/local != \/\u\s\r\/\l\o\c\a\l\/\H\o\m\e\b\r\e\w ]]
++14:54:40 [[ /usr/local/Cellar = \/\u\s\r\/\l\o\c\a\l\/\H\o\m\e\b\r\e\w\/\C\e\l\l\a\r ]]
++14:54:40 source /usr/local/Homebrew/Library/Homebrew/utils/analytics.sh
++14:54:40 setup-analytics
++14:54:40 local git_config_file=/usr/local/Homebrew/.git/config
++14:54:40 migrate-legacy-uuid-file
++14:54:40 [[ -n '' ]]
+++14:54:40 git config --file=/usr/local/Homebrew/.git/config --get homebrew.analyticsmessage
++14:54:40 local message_seen=true
+++14:54:40 git config --file=/usr/local/Homebrew/.git/config --get homebrew.analyticsdisabled
++14:54:40 local analytics_disabled=
++14:54:40 [[ true != \t\r\u\e ]]
++14:54:40 [[ '' = \t\r\u\e ]]
+++14:54:40 git config --file=/usr/local/Homebrew/.git/config --get homebrew.analyticsuuid
++14:54:40 [[ -n '' ]]
++14:54:40 export HOMEBREW_ANALYTICS_ID
++14:54:40 export HOMEBREW_ANALYTICS_USER_UUID
++14:54:40 report-analytics-screenview-command
++14:54:40 [[ -n '' ]]
++14:54:40 [[ -n '' ]]
++14:54:40 [[ 1 != 1 ]]
++14:54:40 [[ list = \b\u\n\d\l\e ]]
++14:54:40 [[ list = \s\e\r\v\i\c\e\s ]]
++14:54:40 [[ -f /usr/local/Homebrew/Library/Homebrew/cmd/list.rb ]]
++14:54:40 case "$HOMEBREW_COMMAND" in
++14:54:40 args=(--max-time 3 --user-agent "$HOMEBREW_USER_AGENT_CURL" --data v=1 --data aip=1 --data t=screenview --data tid="$HOMEBREW_ANALYTICS_ID" --data cid="$HOMEBREW_ANALYTICS_USER_UUID" --data an="$HOMEBREW_PRODUCT" --data av="$HOMEBREW_VERSION" --data cd="$HOMEBREW_COMMAND")
++14:54:40 local args
++14:54:40 [[ -z '' ]]
++14:54:40 disown
++14:54:40 [[ -n '' ]]
++14:54:40 source /usr/local/Homebrew/Library/Homebrew/utils/ruby.sh
++14:54:40 setup-ruby-path
++14:54:40 local vendor_dir
++14:54:40 local vendor_ruby_current_version
++14:54:40 local vendor_ruby_path
++14:54:40 local ruby_version_major
++14:54:40 vendor_dir=/usr/local/Homebrew/Library/Homebrew/vendor
++14:54:40 vendor_ruby_current_version=/usr/local/Homebrew/Library/Homebrew/vendor/portable-ruby/current
++14:54:40 vendor_ruby_path=/usr/local/Homebrew/Library/Homebrew/vendor/portable-ruby/current/bin/ruby
++14:54:40 [[ -z '' ]]
++14:54:40 unset HOMEBREW_RUBY_PATH
++14:54:40 [[ -z '' ]]
++14:54:40 [[ list != \v\e\n\d\o\r\-\i\n\s\t\a\l\l ]]
++14:54:40 [[ -x /usr/local/Homebrew/Library/Homebrew/vendor/portable-ruby/current/bin/ruby ]]
++14:54:40 [[ -n 1 ]]
++14:54:40 HOMEBREW_RUBY_PATH=/System/Library/Frameworks/Ruby.framework/Versions/Current/usr/bin/ruby
++14:54:40 [[ -n /System/Library/Frameworks/Ruby.framework/Versions/Current/usr/bin/ruby ]]
+++14:54:40 /System/Library/Frameworks/Ruby.framework/Versions/Current/usr/bin/ruby --version
++14:54:40 ruby_version_major='ruby 2.0.0p648 (2015-12-16 revision 53162) [universal.x86_64-darwin15]'
++14:54:40 ruby_version_major='2.0.0p648 (2015-12-16 revision 53162) [universal.x86_64-darwin15]'
++14:54:40 ruby_version_major=2
++14:54:40 [[ 2 != \2 ]]
++14:54:40 [[ -n '' ]]
++14:54:40 export HOMEBREW_RUBY_PATH
++14:54:40 [[ 1 -gt 0 ]]
++14:54:40 set -- list
++14:54:40 update-preinstall
++14:54:40 [[ -z '' ]]
++14:54:40 [[ -z '' ]]
++14:54:40 [[ -z '' ]]
++14:54:40 export HOMEBREW_ENABLE_AUTO_UPDATE_MIGRATION=1
++14:54:40 HOMEBREW_ENABLE_AUTO_UPDATE_MIGRATION=1
++14:54:40 [[ list = \i\n\s\t\a\l\l ]]
++14:54:40 [[ list = \u\p\g\r\a\d\e ]]
++14:54:40 [[ list = \t\a\p ]]
++14:54:40 [[ /usr/local/Homebrew = \/\u\s\r\/\l\o\c\a\l ]]
++14:54:40 export HOMEBREW_NO_AUTO_UPDATE=1
++14:54:40 HOMEBREW_NO_AUTO_UPDATE=1
++14:54:40 exec /System/Library/Frameworks/Ruby.framework/Versions/Current/usr/bin/ruby -W0 /usr/local/Homebrew/Library/Homebrew/brew.rb list
alembic
ant
autoconf
automake
awscli
...
brew list > brew_list_output 2>&1  0.32s user 0.15s system 103% cpu 0.460 total

/System/Library/Frameworks/Ruby.framework/Versions/Current/usr/bin/ruby -W0 /usr/local/Homebrew/Library/Homebrew/brew.rb list
real	0m0.248s
user	0m0.228s
sys	0m0.044s
```


brew cellar:
```
+ set -x
+ PS4='+\t '
+14:55:06 set +o posix
++14:55:06 quiet_cd /usr/local/bin/
++14:55:06 cd /usr/local/bin/
++14:55:06 pwd -P
+14:55:06 BREW_FILE_DIRECTORY=/usr/local/bin
+14:55:06 HOMEBREW_BREW_FILE=/usr/local/bin/brew
+14:55:06 HOMEBREW_PREFIX=/usr/local
+14:55:06 [[ -z /usr/local ]]
+14:55:06 [[ /usr/local = \/\u\s\r\/\l\o\c\a\l\/\b\i\n\/\b\r\e\w ]]
+14:55:06 HOMEBREW_REPOSITORY=/usr/local
+14:55:06 [[ -L /usr/local/bin/brew ]]
++14:55:06 symlink_target_directory /usr/local/bin/brew /usr/local/bin
+++14:55:06 readlink /usr/local/bin/brew
++14:55:06 local target=../Homebrew/bin/brew
+++14:55:06 dirname ../Homebrew/bin/brew
++14:55:06 local target_dirname=../Homebrew/bin
++14:55:06 local directory=/usr/local/bin
++14:55:06 quiet_cd /usr/local/bin
++14:55:06 cd /usr/local/bin
++14:55:06 quiet_cd ../Homebrew/bin
++14:55:06 cd ../Homebrew/bin
++14:55:06 pwd -P
+14:55:06 BREW_FILE_DIRECTORY=/usr/local/Homebrew/bin
+14:55:06 HOMEBREW_REPOSITORY=/usr/local/Homebrew
+14:55:06 [[ -L /usr/local/bin/brew ]]
++14:55:06 symlink_target_directory /usr/local/bin/brew /usr/local/bin
+++14:55:06 readlink /usr/local/bin/brew
++14:55:06 local target=../Homebrew/bin/brew
+++14:55:06 dirname ../Homebrew/bin/brew
++14:55:06 local target_dirname=../Homebrew/bin
++14:55:06 local directory=/usr/local/bin
++14:55:06 quiet_cd /usr/local/bin
++14:55:06 cd /usr/local/bin
++14:55:06 quiet_cd ../Homebrew/bin
++14:55:06 cd ../Homebrew/bin
++14:55:06 pwd -P
+14:55:06 USR_LOCAL_BREW_FILE_DIRECTORY=/usr/local/Homebrew/bin
+14:55:06 USR_LOCAL_HOMEBREW_REPOSITORY=/usr/local/Homebrew
+14:55:06 [[ /usr/local/Homebrew = \/\u\s\r\/\l\o\c\a\l\/\H\o\m\e\b\r\e\w ]]
+14:55:06 HOMEBREW_PREFIX=/usr/local
+14:55:06 HOMEBREW_LIBRARY=/usr/local/Homebrew/Library
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z vim ]]
+14:55:06 VAR_NEW=HOMEBREW_EDITOR
+14:55:06 [[ -n '' ]]
+14:55:06 export HOMEBREW_EDITOR=vim
+14:55:06 HOMEBREW_EDITOR=vim
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 VAR_NEW=HOMEBREW_PATH
+14:55:06 [[ -n '' ]]
+14:55:06 for VAR in AWS_ACCESS_KEY_ID AWS_SECRET_ACCESS_KEY BINTRAY_USER BINTRAY_KEY BROWSER EDITOR GIT PATH VISUAL
+14:55:06 [[ -z '' ]]
+14:55:06 continue
+14:55:06 [[ -n '' ]]
+14:55:06 source /usr/local/Homebrew/Library/Homebrew/brew.sh
+++14:55:06 git -C /usr/local/Homebrew describe --tags --dirty
++14:55:06 HOMEBREW_VERSION=1.2.1-96-gfb33acb-dirty
++14:55:06 HOMEBREW_USER_AGENT_VERSION=1.2.1-96-gfb33acb-dirty
++14:55:06 [[ -z 1.2.1-96-gfb33acb-dirty ]]
++14:55:06 export HOMEBREW_COMMAND_DEPTH=1
++14:55:06 HOMEBREW_COMMAND_DEPTH=1
+++14:55:06 locale charmap
++14:55:06 [[ UTF-8 != \U\T\F\-\8 ]]
++14:55:06 [[ -d /usr/local/Homebrew/Cellar ]]
++14:55:06 HOMEBREW_CELLAR=/usr/local/Cellar
++14:55:06 case "$*" in
++14:55:06 echo /usr/local/Cellar
/usr/local/Cellar
++14:55:06 exit 0

brew --cellar > brew_cellar_output 2>&1  0.02s user 0.02s system 95% cpu 0.047 total
```