@kenden 2017-06-13
[calling the aws autocomplete](https://github.com/robbyrussell/oh-my-zsh/blob/master/plugins/aws/aws.plugin.zsh#L31) is also slow, because it also uses brew.
Could we replace with this?
`_aws_zsh_completer_path=/usr/local/opt/awscli/libexec/bin/aws_zsh_completer.sh`
This is from this PR: https://github.com/robbyrussell/oh-my-zsh/pull/4886